+++ packages/core-js/src/core/config.ts
@@ -6,14 +6,16 @@ export const stagingEnvConfig: EnvConfig = {
   bffBaseUrl: 'https://apigw.oktostage.com',
   paymasterAddress: Constants.ENV_CONFIG.STAGING.PAYMASTER_ADDRESS,
   jobManagerAddress: Constants.ENV_CONFIG.STAGING.JOB_MANAGER_ADDRESS,
+  entryPointAddress: Constants.ENV_CONFIG.STAGING.ENTRYPOINT_CONTRACT_ADDRESS,
   chainId: Constants.ENV_CONFIG.STAGING.CHAIN_ID,
 };
 
 export const sandboxEnvConfig: EnvConfig = {
   gatewayBaseUrl: 'https://sandbox-okto-gateway.oktostage.com',
-  bffBaseUrl: 'https://3p-bff.oktostage.com',
+  bffBaseUrl: 'https://sandbox-api.okto.tech',
   paymasterAddress: Constants.ENV_CONFIG.SANDBOX.PAYMASTER_ADDRESS,
   jobManagerAddress: Constants.ENV_CONFIG.SANDBOX.JOB_MANAGER_ADDRESS,
+  entryPointAddress: Constants.ENV_CONFIG.SANDBOX.ENTRYPOINT_CONTRACT_ADDRESS,
   chainId: Constants.ENV_CONFIG.SANDBOX.CHAIN_ID,
 };