+++ packages/react-native/src/core/index.ts
@@ -3,47 +3,41 @@ import {
   OktoClient as OktoCoreClient,
   type OktoClientConfig,
 } from '@okto_web3/core-js-sdk';
+import type { SessionConfig } from '@okto_web3/core-js-sdk/core';
 import type { RpcError } from '@okto_web3/core-js-sdk/errors';
 import type { Address, AuthData } from '@okto_web3/core-js-sdk/types';
-import type { SessionConfig } from '@okto_web3/core-js-sdk/core';
-import { decryptData, encryptData } from '../utils/encryptionUtils.js';
-import { setStorage, getStorage, clearStorage } from '../utils/storageUtils.js';
+import { clearStorage, getStorage, setStorage } from '../utils/storageUtils.js';
 
 class OktoClient extends OktoCoreClient {
-  private _clientPrivateKey: string;
-
   constructor(config: OktoClientConfig) {
     super(config);
-    this._clientPrivateKey = config.clientPrivateKey;
-    this.initializeSessionConfig();
+    this.initializeSession();
   }
 
-  private initializeSessionConfig(): void {
-    const encryptedSession = getStorage('session');
-    if (encryptedSession) {
-      const sessionConfig = decryptData<SessionConfig>(
-        encryptedSession,
-        this._clientPrivateKey,
-      );
-      if (sessionConfig) {
-        this.setSessionConfig(sessionConfig);
-      }
+  private initializeSession(): void {
+    const session = getStorage('okto_session');
+    if (session) {
+      this.setSessionConfig(JSON.parse(session));
+      this.syncUserKeys();
     }
   }
 
   override loginUsingOAuth(
     data: AuthData,
+    onSuccess?: (session: SessionConfig) => void,
   ): Promise<Address | RpcError | undefined> {
     return super.loginUsingOAuth(data, (session) => {
-      setStorage('session', encryptData(session, this._clientPrivateKey));
+      setStorage('okto_session', JSON.stringify(session));
       this.setSessionConfig(session);
+      onSuccess?.(session);
     });
   }
 
   override sessionClear(): void {
-    clearStorage();
+    clearStorage('okto_session');
     return super.sessionClear();
   }
 }
 
 export { OktoClient };
+export type { OktoClientConfig };