+++ packages/react-native/src/index.ts
@@ -1,6 +1,8 @@
 import 'react-native-get-random-values';
 
 export { OktoProvider } from './context/OktoProvider.js';
+export { OktoClient } from './core/index.js';
+export type { OktoClientConfig } from './core/index.js';
 export { useOkto } from './hooks/index.js';
 
 export * from './explorer/index.js';