+++ packages/core-js/CHANGELOG.md
@@ -1,5 +1,54 @@
 # @okto_web3/core-js-sdk
 
+## 0.6.0
+
+### Minor Changes
+
+- 24248fa: update job manager abi
+- bd0cc04: Update signMessage Payload Generation
+
+## 0.5.1
+
+### Patch Changes
+
+- add getUserOperationGasPrice call
+
+## 0.5.0
+
+### Minor Changes
+
+- d2845ef: add support for signMessage and signTypedData
+
+## 0.4.0
+
+### Minor Changes
+
+- fix local storage
+
+## 0.3.1
+
+### Patch Changes
+
+- 2d3f0ae: add better session management
+
+## 0.3.0
+
+### Minor Changes
+
+- f2c51c4: entrypoint contract update
+
+## 0.2.1
+
+### Patch Changes
+
+- 28a93bf: add readme
+
+## 0.2.0
+
+### Minor Changes
+
+- validation fix
+
 ## 0.1.0
 
 ### Minor Changes