+++ packages/react-native/CHANGELOG.md
@@ -1,5 +1,82 @@
 # @okto_web3/react-native-sdk
 
+## 0.5.0
+
+### Minor Changes
+
+- 24248fa: update job manager abi
+- bd0cc04: Update signMessage Payload Generation
+
+### Patch Changes
+
+- Updated dependencies [24248fa]
+- Updated dependencies [bd0cc04]
+  - @okto_web3/core-js-sdk@0.6.0
+
+## 0.4.1
+
+### Patch Changes
+
+- add getUserOperationGasPrice call
+- Updated dependencies
+  - @okto_web3/core-js-sdk@0.5.1
+
+## 0.4.0
+
+### Minor Changes
+
+- d2845ef: add support for signMessage and signTypedData
+
+### Patch Changes
+
+- Updated dependencies [d2845ef]
+  - @okto_web3/core-js-sdk@0.5.0
+
+## 0.3.0
+
+### Minor Changes
+
+- fix local storage
+
+### Patch Changes
+
+- Updated dependencies
+  - @okto_web3/core-js-sdk@0.4.0
+
+## 0.2.1
+
+### Patch Changes
+
+- 2d3f0ae: add better session management
+- Updated dependencies [2d3f0ae]
+  - @okto_web3/core-js-sdk@0.3.1
+
+## 0.2.0
+
+### Minor Changes
+
+- f2c51c4: entrypoint contract update
+
+### Patch Changes
+
+- Updated dependencies [f2c51c4]
+  - @okto_web3/core-js-sdk@0.3.0
+
+## 0.1.2
+
+### Patch Changes
+
+- 28a93bf: add readme
+- Updated dependencies [28a93bf]
+  - @okto_web3/core-js-sdk@0.2.1
+
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @okto_web3/core-js-sdk@0.2.0
+
 ## 0.1.0
 
 ### Minor Changes